Fix/cart (#15)
* :adhesive_bandage: Response수정, 불필요 파일 삭제, 유저검증 추가

* :sparkles: data.sql추가

+++ src/main/java/com/sparta/project/icylattei/cart/service/CartService.java
@@ -11,6 +11,7 @@ import com.sparta.project.icylattei.user.entity.User;
 import com.sparta.project.icylattei.user.repository.UserRepository;
 import com.sparta.project.icylattei.userDetails.UserDetailsImpl;
 import jakarta.transaction.Transactional;
+import java.util.ArrayList;
 import java.util.List;
 import lombok.AllArgsConstructor;
 import lombok.RequiredArgsConstructor;
@@ -25,20 +26,28 @@ public class CartService {
     private final ProductRepository productRepository;
 
     public CartResponseDto createCart(CartRequestDto requestDto, User user) {
-        String productName = requestDto.getProductName();
-        Product product = productRepository.findByProductName(productName);
+        Long productId = requestDto.getProductId();
+        Product product = productRepository.findById(productId)
+            .orElseThrow(() -> new IllegalArgumentException("상품을 찾을 수 없습니다."));
         Cart cart = cartRepository.save(new Cart(product, requestDto, user));
         return new CartResponseDto(cart);
     }
 
     public CartsResponseDto getCart(User user) {
         List<Cart> cart = cartRepository.findAllByUserAndCartStatus(user, "장바구니");
-        return new CartsResponseDto(cart);
+        List<CartResponseDto> cartRespones = new ArrayList<>();
+        for (Cart value : cart) {
+            cartRespones.add(new CartResponseDto(value));
+        }
+        return new CartsResponseDto(cartRespones);
     }
 
     public CartResponseDto deleteCart(Long cartId, User user) {
         Cart cart = cartRepository.findById(cartId)
             .orElseThrow(() -> new IllegalArgumentException("장바구니를 찾을 수 없습니다."));
+        if(!cart.getUser().getId().equals(user.getId())){
+            throw new IllegalArgumentException("수정할 수 없습니다.");
+        }
         cartRepository.delete(cart);
         return new CartResponseDto(cart);
     }
@@ -47,9 +56,11 @@ public class CartService {
     public CartResponseDto updateCart(Long cartId, CartRequestDto requestDto, User user) {
         Cart cart = cartRepository.findById(cartId)
             .orElseThrow(() -> new IllegalArgumentException("장바구니를 찾을 수 없습니다."));
-
-        String productName = requestDto.getProductName();
-        Product product = productRepository.findByProductName(productName);
+        if(!cart.getUser().getId().equals(user.getId())){
+            throw new IllegalArgumentException("수정할 수 없습니다.");
+        }
+        Long productId = requestDto.getProductId();
+        Product product = productRepository.findById(productId).orElseThrow(() -> new IllegalArgumentException("상품을 찾을 수 없습니다."));
         cart.update(product, requestDto);
         return new CartResponseDto(cart);
     }
